+++ recipe/drupal7.php
@@ -24,7 +24,6 @@ task('deploy', [
 //Set Drupal 7 site. Change if you use different site
 set('drupal_site', 'default');
 
-
 //Drupal 7 shared dirs
 set('shared_dirs', [
     'sites/{{drupal_site}}/files',
@@ -44,7 +43,9 @@ set('writable_dirs', [
 //Create and upload Drupal 7 settings.php using values from secrets
 task('drupal:settings', function () {
     if (askConfirmation('Are you sure to generate and upload settings.php file?')) {
-        $basepath = dirname(__FILE__) . '/drupal7';
+
+        //Get template
+        $template = get('settings_template');
 
         //Import secrets
         $secrets = get('settings');
@@ -66,13 +67,13 @@ task('drupal:settings', function () {
         }
 
         //Create settings from template
-        $settings = file_get_contents($basepath . '/settings.php');
+        $settings = file_get_contents($template);
 
         $settings = strtr($settings, $replacements);
 
         writeln('settings.php created succesfuly');
 
-        $tmpFilename = tempnam($basepath, 'tmp_settings_');
+        $tmpFilename = tempnam(sys_get_temp_dir(), 'tmp_settings_');
         file_put_contents($tmpFilename, $settings);
 
         upload($tmpFilename, '{{deploy_path}}/shared/sites/{{drupal_site}}/settings.php');